refactor : refactor realtime module as per design principles
